requirements.txt
Django>=3.0.0
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
django-filter==2.2.0
dynamic-rest==1.9.6
mysqlclient==1.4.4
ujson==1.35
requests==2.22.0
django-storages[google]==1.8
gunicorn==20.0.4
django-widget-tweaks==1.4.5
sorl-thumbnail==12.6.2
Pillow==7.0.0


# Test, Coverage, Standard
flake8==3.7.8
flake8-bugbear==19.8.0
flake8-import-order==0.18.1
flake8-builtins==1.4.1
flake8-docstrings==1.5.0
pep8-naming==0.8.2
coverage==4.5.4
unittest-xml-reporting==2.5.1
